Add files via upload
Project 1 CHADA clock CS210 
This project was creating a clock that allowed the user to add 1 hr, 	 1 min, or 1 sec to its pulled time.  During the project I really felt like I struggled with breaking the code into functions because I was ultimately confused with C++ as a whole.  I did not understand it one bit, managed to complete with minimal code in project1.cpp. However, with this said I feel there was some issues in the final presentation of my project that could have benefitted from minor changes in the user display. Along with this I feel there was probably an easier or more efficent way of converting the system time into changeable variables, although I personally feel my method did serve its purpose well.  The most challenging part of this code for me was integrating the system time into individual integers for time and with this I had problems coming up with a method to do so and had to use a lot of the online tutoring, which to me was a little more tough to resolve these complications I went to online tutoring and considering all the tutors I went through all had their own opinions or different way of teaching.  As well looking at my own personal text book to find ways to pull individual factors from a string as well convert a string to an integer resulting in the method I used in this code, but strings have always been a huge weakness through the entire duration of this course.  The skills from this project that will be particularly transferrable to other course work are the functions involved in this code. This is due the fact that throughout the entirety of programming in the functions have had a strong role in creating efficient and well working code and with this any practice of functions proves beneficial. To make this program maintainable, readable, and adaptable I used functions as well as whitespace to separate code into sections as well prevent my code from becoming cluttered thus enhancing readability.  Overall never really understanding when everyone that helped me meant by using OOP or “treat it as OOP” practices.